refactor: decouple sync engine, HTTP client, and metadata XML for testability
Extract the sync algorithm into MavenSyncEngine taking injected repositories
and SyncOptions. Split MavenHttpClient construction so an HttpClient (or
MockEngine in tests) can be injected, with HttpClientConfig holding tunables.
Move maven-metadata.xml render/parse into MavenMetadataXml +
MavenMetadataXmlParser, and the recursive crawl into MavenRepositoryCrawler
exposing a Flow<ArtifactMetadata>. Replace the hardcoded lastUpdated
timestamp with a kotlinx-datetime Clock seam (defaults to system clock).
Surface trustAllCerts in RepositoryConfig, defaulted to true to preserve
current behavior.
