+++ gradle/libs.versions.toml
@@ -13,6 +13,7 @@ ktfmt = { id = "com.ncorti.ktfmt.gradle", version = "0.26.0" }
 
 [libraries]
 kotlinx-coroutines = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version = "1.10.2" }
+kotlinx-datetime = { module = "org.jetbrains.kotlinx:kotlinx-datetime", version = "0.7.0" }
 oshai-kotlinLogging = { module = "io.github.oshai:kotlin-logging-jvm", version = "7.0.7" }
 pearx-kasechange = { module = "net.pearx.kasechange:kasechange-jvm", version = "1.4.1" }
 guava = { module = "com.google.guava:guava", version = "33.4.8-jre" }
